pyproject.toml
[tool.poetry]
name = "emails-parser"
version = "0.1.0"
description = "Email transaction parser with Google OAuth and Gmail API integration"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
google-auth = "^2.23.4"
google-auth-oauthlib = "^1.1.0"
google-api-python-client = "^2.108.0"
beautifulsoup4 = "^4.12.2"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
requests = "^2.31.0"
lxml = "^4.9.3"
google-generativeai = "^0.3.0"
selenium = "4.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"